feat: Add Rmd running NEArender
It's a redacted and simplified version of the code used in 4vs2 project.
Majority of the changes are due to not having 4 AGS sets (one for each
cancer) but only 1 (genes common to all 4 cancers).
Apart from this, the FGS and network objects were loaded from RData,
thus no need to create them.
The output is also greatly simplified.